+++ qiskit/pulse/__init__.py
@@ -127,7 +127,7 @@ Exceptions
 from .channels import (DriveChannel, MeasureChannel, AcquireChannel,
                        ControlChannel, RegisterSlot, MemorySlot)
 from .cmd_def import CmdDef
-from .commands import (Instruction, Acquire, AcquireInstruction, FrameChange,
+from .commands import (Acquire, AcquireInstruction, FrameChange,
                        PersistentValue, SamplePulse, Snapshot, Kernel,
                        Discriminator, Delay, ParametricPulse,
                        ParametricInstruction, Gaussian,
@@ -135,5 +135,6 @@ from .commands import (Instruction, Acquire, AcquireInstruction, FrameChange,
 from .configuration import LoConfig, LoRange
 from .exceptions import PulseError
 from .instruction_schedule_map import InstructionScheduleMap
+from .instructions import Instruction
 from .interfaces import ScheduleComponent
 from .schedule import Schedule
